% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adrs_sat}
\alias{adrs_sat}
\title{Binary outcome data from single arm trial}
\format{
A data frame with 500 rows and 5 columns:
\describe{
\item{USUBJID}{Unique subject identifiers for patients.}
\item{ARM}{Assigned treatment arm.}
\item{AVAL}{Analysis value, in this dataset an indicator of response.}
\item{PARAM}{Parameter type of \code{AVAL}.}
\item{RESPONSE}{Indicator of response.}
}
}
\usage{
adrs_sat
}
\description{
Binary outcome data from single arm trial
}
\seealso{
Other unanchored datasets: 
\code{\link{adsl_sat}},
\code{\link{adtte_sat}},
\code{\link{agd}},
\code{\link{centered_ipd_sat}},
\code{\link{pseudo_ipd_sat}},
\code{\link{weighted_sat}}
}
\concept{unanchored datasets}
\keyword{dataset}
