% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_unanchored.R
\docType{data}
\name{adsl_sat}
\alias{adsl_sat}
\title{Dummy Single-arm Trial Data for Unanchored MAIC}
\format{
A data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}
}
\source{
These datasets are generated by the dummy_unanchored.R script.
}
\usage{
adsl_sat
}
\description{
These datasets contains adsl_sat, adtte_sat, adrs_sat,
agd_sat (AgD of effect modifiers), pseudo_ipd_sat (AgD, tte data)
}
\keyword{dataset}
