% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl_twt}
\alias{adsl_twt}
\title{Patient data from two arm trial}
\format{
A data frame with 1000 rows and 9 columns:
\describe{
\item{X}{The ID of the patients.}
\item{USUBJID}{Unique subject identifiers for patients.}
\item{ARM}{Assigned treatment arm.}
\item{AGE}{Age in years at baseline.}
\item{SEX}{Sex of patient recorded as character "Male"/"Female"}
\item{SMOKE}{Smoking status at baseline as integer 0/1}
\item{ECOG0}{ECOG score at baseline as integer 0/1}
\item{N_PR_THER}{Number of prior therapies received as integer 1, 2, 3, 4}
\item{SEX_MALE}{Indicator of SEX == "Male" as numeric 1/0}
}
}
\usage{
adsl_twt
}
\description{
Patient data from two arm trial
}
\seealso{
Other anchored datasets: 
\code{\link{adtte_twt}},
\code{\link{agd}},
\code{\link{centered_ipd_twt}},
\code{\link{pseudo_ipd_twt}}
}
\concept{anchored datasets}
\keyword{dataset}
