% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adtte_sat}
\alias{adtte_sat}
\title{Survival data from single arm trial}
\format{
A data frame with 500 rows and 10 columns:
\describe{
\item{USUBJID}{}
\item{ARM}{}
\item{EVNT}{}
\item{AVAL}{}
\item{AVALU}{}
\item{PARAMCD}{}
\item{PARAM}{}
\item{CNSR}{}
\item{TIME}{}
\item{EVENT}{}
}
}
\usage{
adtte_sat
}
\description{
Survival data from single arm trial
}
\seealso{
Other unanchored datasets: 
\code{\link{adrs_sat}},
\code{\link{adsl_sat}},
\code{\link{agd}},
\code{\link{centered_ipd_sat}},
\code{\link{pseudo_ipd_sat}}
}
\concept{unanchored datasets}
\keyword{dataset}
