% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adtte_twt}
\alias{adtte_twt}
\title{Survival data from two arm trial}
\format{
A data frame with 1000 rows and 10 columns:
\describe{
\item{USUBJID}{}
\item{ARM}{}
\item{EVNT}{}
\item{AVAL}{}
\item{AVALU}{}
\item{PARAMCD}{}
\item{PARAM}{}
\item{CNSR}{}
\item{TIME}{}
\item{EVENT}{}
}
}
\usage{
adtte_twt
}
\description{
Survival data from two arm trial
}
\seealso{
Other anchored datasets: 
\code{\link{adsl_twt}},
\code{\link{agd}},
\code{\link{centered_ipd_twt}},
\code{\link{pseudo_ipd_twt}}
}
\concept{anchored datasets}
\keyword{dataset}
