% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agd}
\alias{agd}
\title{Aggregate effect modifier data from published study}
\format{
A data frame with 3 rows and 9 columns:
\describe{
\item{STUDY}{}
\item{ARM}{}
\item{N}{}
\item{AGE_MEAN}{}
\item{AGE_MEDIAN}{}
\item{AGE_SD}{}
\item{SEX_MALE_COUNT}{}
\item{ECOG0_COUNT}{}
\item{SMOKE_COUNT}{}
}
}
\usage{
agd
}
\description{
This data is formatted to be used in \code{\link[=center_ipd]{center_ipd()}}.
}
\seealso{
Other unanchored datasets: 
\code{\link{adrs_sat}},
\code{\link{adsl_sat}},
\code{\link{adtte_sat}},
\code{\link{centered_ipd_sat}},
\code{\link{pseudo_ipd_sat}}

Other anchored datasets: 
\code{\link{adsl_twt}},
\code{\link{adtte_twt}},
\code{\link{centered_ipd_twt}},
\code{\link{pseudo_ipd_twt}}
}
\concept{anchored datasets}
\concept{unanchored datasets}
\keyword{dataset}
