% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anchored-data}
\alias{anchored-data}
\alias{adsl_twt}
\alias{adtte_twt}
\alias{pseudo_ipd_twt}
\alias{centered_ipd_twt}
\title{Dummy Trial Data for anchored MAIC}
\format{
\strong{adsl_twt}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

\strong{adtte_twt}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

\strong{pseudo_ipd_twt}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

\strong{centered_ipd_twt}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

An object of class \code{data.frame} with 1000 rows and 10 columns.

An object of class \code{data.frame} with 800 rows and 3 columns.

An object of class \code{data.frame} with 1000 rows and 15 columns.
}
\source{
These datasets are generated by the dummy_anchored.R script.
}
\usage{
adsl_twt

adtte_twt

pseudo_ipd_twt

centered_ipd_twt
}
\description{
These datasets contains adsl_twt, adtte_twt,
centered_ipd_twt,  pseudo_ipd_twt (AgD, tte data)
}
\keyword{dataset}
\keyword{datasets}
