% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_km.R
\name{basic_kmplot}
\alias{basic_kmplot}
\title{Basic Kaplan Meier (KM) plot function}
\usage{
basic_kmplot(
  kmdat,
  time_scale,
  show_risk_set = TRUE,
  main_title = "Kaplan-Meier Curves",
  subplot_heights = NULL,
  suppress_plot_layout = FALSE,
  use_colors = NULL,
  use_line_types = NULL
)
}
\arguments{
\item{kmdat}{a data.frame, must consist 'treatment', 'time' (unit in days), 'n.risk', 'censor', 'surv', similar to an output from \code{maicplus:::survfit_makeup}}

\item{time_scale}{a string, time unit of median survival time, taking a value of 'year', 'month', 'week' or 'day'}

\item{show_risk_set}{logical, show risk set table or not, TRUE by default}

\item{main_title}{a string, main title of the KM plot}

\item{subplot_heights}{a numeric vector, heights argument to \code{graphic::layout()},NULL by default which means user will use the default setting}

\item{suppress_plot_layout}{logical, suppress the layout setting in this function so that user can specify layout outside of the function, FALSE by default}

\item{use_colors}{a character vector of length up to 3, colors to the KM curves, it will be passed to 'col' of \code{lines()}}

\item{use_line_types}{a numeric vector of length up to 3, line type to the KM curves, it will be passed to 'lty' of \code{lines()}}
}
\value{
a KM plot with or without risk set table appended at the bottom, with up to 3 KM curves
}
\description{
This function can generate a basic KM plot with or without risk set table appended at the bottom.
In a single plot, it can include up to 3 KM curves. This depends on number of levels in 'treatment' column in the input data.frame \code{kmdat}
}
