% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_HR}
\alias{bootstrap_HR}
\title{Bootstrapping for MAIC weighted hazard ratios}
\usage{
bootstrap_HR(
  internal,
  i,
  centered_colnames,
  external,
  model,
  min_weight = 1e-04
)
}
\arguments{
\item{internal}{A data frame containing individual patient data
from the internal IPD study.}

\item{i}{Index used to select a sample within \code{\link{boot}}.}

\item{centered_colnames}{A character vector giving the names of the covariates to use
in matching. These names must match the column names in internal dataset.}

\item{external}{A data frame containing pseudo individual patient data
from the comparator study needed to derive the relative treatment effect.
The outcome variables names must match intervention_data.}

\item{model}{A model formula in the form 'Surv(Time, Event==1) ~ ARM'.
Variable names need to match the corresponding columns in intervention_data.}

\item{min_weight}{A numeric value that defines the minimum weight allowed.
This value (default 0.0001) will replace weights estimated at 0 in a sample.}
}
\value{
The HR as a numeric value.
}
\description{
Bootstrapping for MAIC weighted hazard ratios
}
\details{
This function is intended to be used in conjunction with the
\code{\link{boot}} function to return the statistic to be
bootstrapped. In this case by performing MAIC weighting using
{\link{estimate_weights}} and returning a weighted hazard ratio (HR) from a
Cox proportional hazards model. This is used as the 'statistic' argument in
the boot function.
}
