% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{calculate_weights_legend}
\alias{calculate_weights_legend}
\title{Calculate Statistics for Weight Plot Legend}
\usage{
calculate_weights_legend(weighted_data)
}
\arguments{
\item{weighted_data}{object returned after calculating weights using \link{estimate_weights}}
}
\value{
list of ESS, ESS reduction, median value of scaled and unscaled weights, and missing count
}
\description{
Calculates ESS reduction and median weights which is used to create legend for weights plot
}
\examples{
\dontrun{
load(system.file("extdata", "weighted_data.rda", package = "maicplus", mustWork = TRUE))
calculate_weights_legend(weighted_data)
}
}
\keyword{internal}
