% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centered_ipd_sat}
\alias{centered_ipd_sat}
\title{Centered patient data from single arm trial}
\format{
A data frame with 500 rows and 14 columns:
\describe{
\item{USUBJID}{Unique subject identifiers for patients.}
\item{ARM}{Assigned treatment arm.}
\item{AGE}{Age in years at baseline.}
\item{SEX}{Sex of patient recorded as character \code{"Male"}/\code{"Female"}.}
\item{SMOKE}{Smoking status at baseline as integer \code{0}/\code{1}.}
\item{ECOG0}{ECOG score at baseline as integer \code{0}/\code{1}.}
\item{N_PR_THER}{Number of prior therapies received as integer \verb{1, 2, 3, 4}.}
\item{SEX_MALE}{Indicator of \code{SEX == "Male"} as numeric \code{1}/\code{0}.}
\item{AGE_CENTERED}{Age in years at baseline relative to average in aggregate data \link{agd}.}
\item{AGE_MEDIAN_CENTERED}{Age greater/less than median age in \link{agd} coded as  TODO .}
\item{AGE_SQUARED_CENTERED}{TODO}
\item{SEX_MALE_CENTERED}{TODO}
\item{ECOG0_CENTERED}{TODO}
\item{SMOKE_CENTERED}{TODO}
}
}
\usage{
centered_ipd_sat
}
\description{
Centered patient data from single arm trial
}
\seealso{
Other unanchored datasets: 
\code{\link{adrs_sat}},
\code{\link{adsl_sat}},
\code{\link{adtte_sat}},
\code{\link{agd}},
\code{\link{pseudo_ipd_sat}},
\code{\link{weighted_sat}}
}
\concept{unanchored datasets}
\keyword{dataset}
