% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{check_weights}
\alias{check_weights}
\title{Check to see if weights are optimized correctly}
\usage{
check_weights(
  optimized,
  processed_agd,
  mean_digits = 2,
  prop_digits = 2,
  sd_digits = 3
)
}
\arguments{
\item{optimized}{object returned after calculating weights using \code{\link{estimate_weights}}}

\item{processed_agd}{a data frame, object returned after using \code{\link{process_agd}} or aggregated data following the same naming convention}

\item{digits}{number of digits for rounding summary table}
}
\value{
data.frame of weighted and unweighted covariate averages of the IPD
}
\description{
This function checks to see if the optimization is done properly by checking the covariate averages
before and after adjustment.
}
