% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{complete.agd}
\alias{complete.agd}
\title{Calculate pooled arm statistics in AgD based on arm-specific statistics}
\usage{
complete.agd(use.agd)
}
\arguments{
\item{use_agd}{}
}
\description{
Calculate pooled arm statistics in AgD based on arm-specific statistics
}
