% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{complete_agd}
\alias{complete_agd}
\title{Calculate pooled arm statistics in Aggregated Data (AgD) based on arm-specific statistics}
\usage{
complete_agd(use_agd)
}
\arguments{
\item{use_agd}{aggregated data that is processed within process_agd}
}
\value{
Complete N, count, mean, sd, and median for the pooled arm
}
\description{
This is a convenient function to pool arm statistics. This function is called
within process_agd and when the ARM is not equal to "Total". Note pooled
median can't be calculated and it is only an approximation.
}
