% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic_analysis.R
\name{dummy_mod}
\alias{dummy_mod}
\title{Derive treatment effect in the analysis step of MAIC}
\usage{
dummy_mod(Ntrt, ntrt, Ncom, ncom, link = poisson(link = "identity"))
}
\arguments{
\item{link}{}
}
\description{
Derive treatment effect in the analysis step of MAIC
}
