% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{ess_footnote_text}
\alias{ess_footnote_text}
\title{Note on Expected Sample Size Reduction}
\usage{
ess_footnote_text(width = 0.9 * getOption("width"))
}
\arguments{
\item{width}{Number of characters to break string into new lines (\verb{\\n}).}
}
\value{
A character string
}
\description{
Note on Expected Sample Size Reduction
}
\keyword{internal}
