% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{estimate_weights}
\alias{estimate_weights}
\title{Derive individual weights in the matching step of MAIC}
\usage{
estimate_weights(data, centered_colnames, start_val = 0, method = "BFGS", ...)
}
\arguments{
\item{data}{a numeric matrix, centered covariates of IPD, no missing value in any cell is allowed}

\item{centered_colnames}{a character or numeric vector (column indicators) of centered covariates}

\item{start_val}{a scalar, the starting value for all coefficients of the propensity score regression}

\item{method}{a string, name of the optimization algorithm (see 'method' argument of \code{base::optim()}).
The default is \code{"BFGS"}, other options are \code{"Nelder-Mead"}, \code{"CG"}, \code{"L-BFGS-B"}, \code{"SANN"}, and \code{"Brent"}}

\item{...}{all other arguments from \code{base::optim()}}
}
\value{
a list with the following 4 elements,
\describe{
\item{data}{a data.frame, includes the input \code{data} with appended column 'weights' and 'scaled_weights'.
Scaled weights has a summation to be the number of rows in \code{data} that has no missing value in any of the
effect modifiers}
\item{centered_colnames}{column names of centered effect modifiers in \code{data}}
\item{nr_missing}{number of rows in \code{data} that has at least 1 missing value in specified centered effect
modifiers}
\item{ess}{effective sample size, square of sum divided by sum of squares}
\item{opt}{R object returned by \code{base::optim()}, for assess convergence and other details}
}
}
\description{
Assuming data is properly processed, this function takes individual patient data (IPD) with centered covariates
(effect modifiers and/or prognostic variables) as input, and generates weights for each individual in IPD trial that
matches the chosen statistics of those covariates in Aggregated Data (AgD) trial.
}
