% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{ext_tte_transfer}
\alias{ext_tte_transfer}
\title{helper function: transform TTE ADaM data to suitable input for survival R pkg}
\usage{
ext_tte_transfer(dd, time_scale = "month", trt = NULL)
}
\arguments{
\item{dd}{data frame, ADTTE read via \code{haven::read_sas()}}

\item{time_scale}{a character string, 'year', 'month', 'week' or 'day', time unit of median survival time}
}
\value{
a data frame can be used as input to \code{\link[survival:Surv]{survival::Surv()}}
}
\description{
helper function: transform TTE ADaM data to suitable input for survival R pkg
}
