% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucher.R
\name{find_SE_fromCI}
\alias{find_SE_fromCI}
\title{Calculate standard error from the reported confidence interval.}
\usage{
find_SE_fromCI(RR = NULL, CI_lower = NULL, CI_upper = NULL, CI_perc = 0.95)
}
\arguments{
\item{RR}{Reported estimate of relative risk (i.e. hazard ratio or odds ratio). Note that this estimate is not logged.}

\item{CI_lower}{Reported lower percentile value of the relative risk}

\item{CI_upper}{Reported upper percentile value of the relative risk}

\item{CI_perc}{Reported percentage of confidence interval reported}
}
\value{
Standard error of the relative risk and log relative risk
}
\description{
Often times in analysis of clinical trials, standard error is not reported and only the confidence interval is reported.
For instance, a given study can report a hazard ratio of 0.70 and 98.22\% confidence interval of 0.55 to 0.90.
This function calculates standard error of the estimate given the reported confidence interval.
}
\examples{
find_SE_fromCI(RR = 0.70, CI_lower = 0.55, CI_upper = 0.90, CI_perc = 0.95)
}
