% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_survival_data}
\alias{generate_survival_data}
\title{Generate time-to-event data}
\usage{
generate_survival_data(
  N = 150,
  ncov = 4,
  lambda = 8.5,
  nu = 1.3,
  censor_rate = 0.96,
  cor = 0.35,
  marginal_mean = 0.6,
  marginal_sd = 0.2,
  beta = rep(-log(0.67), ncov),
  gamma = sort(rep(c(0, -log(0.67)), length = ncov)),
  d0 = log(0.25),
  seed = 1
)
}
\arguments{
\item{N}{sample size}

\item{ncov}{number of covariates to generate}

\item{lambda}{scale parameter in baseline hazard function}

\item{nu}{shape parameter in baseline hazard function}

\item{censor_rate}{rate parameter of the exponential distribution of censoring times}

\item{cor}{correlation between covariates}

\item{marginal_mean}{marginal mean of covariates}

\item{marginal_sd}{marginal standard deviation of covariates}

\item{beta}{vector of coefficients for prognostic factors. Defaults is \code{-log(0.67)} for all covariates.}

\item{gamma}{vector of coefficients for effect modifiers. Default is \code{0} for half of the covariates and
\code{-log(0.67)} for the others.}

\item{d0}{coefficient for the treatment effect for patients with \code{treat = 0}}

\item{seed}{seed used for simulation reproducibility}
}
\value{
data.frame with survival time, censor (event) status, treatment indicator, and covariates
}
\description{
Function generates time-to-event data. We follow the data-generating mechanism from
Bender et al. and Remiro-Azocar et al. 2021 to simulate Weibull-distributed survival times
under a proportional hazards parameterization. As per Remiro-Azocar et al, we set the
default scale parameter (lambda) to 8.5 and shape parameter (nu) to 1.3 to reflect frequently
observed mortality trends in metastatic cancer patients. Censoring times are generated from
exponential distribution where the rate parameter is 0.96 is selected to achieve a censoring rate
of 35\% under the active treatment at baseline (with the values of the covariates set to zero).
}
\references{
Remiro-Azocar et al. 2021
}
