% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-helper.R
\name{get_pseudo_ipd_binary}
\alias{get_pseudo_ipd_binary}
\title{Helper funciton: create pseudo IPD given AgD binary}
\usage{
get_pseudo_ipd_binary(binary_agd, format = c("stacked", "unstacked"))
}
\arguments{
\item{binary_agd}{a data.frame that take different formats depending on \code{format}}

\item{format}{a string, "stacked" or "unstacked"}
}
\value{
a data.frame of pseudo binary IPD, with columns USUBJID, ARM, RESPONSE
}
\description{
Helper funciton: create pseudo IPD given AgD binary
}
\details{
Additional details...
}
