% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-helper.R
\name{get_time_as}
\alias{get_time_as}
\title{Convert Time Values Using Scaling Factors}
\usage{
get_time_as(times, as = NULL)
}
\arguments{
\item{times}{Numeric time values}

\item{as}{A time scale to convert to. One of "days", "weeks", "months", "years"}
}
\value{
Returns a numeric vector calculated from \code{times / get_time_conversion(factor = as)}
}
\description{
Convert Time Values Using Scaling Factors
}
\examples{
get_time_as(50, as = "months")
}
