% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helper.R
\name{km_plot2}
\alias{km_plot2}
\title{Function to plot Kaplan-Meier curves using survminer package}
\usage{
km_plot2(
  combined_data,
  trt,
  trt_ext,
  trt_common = NULL,
  break.x.by = 60,
  endpoint_name = "Overall survival"
)
}
\arguments{
\item{combined_data}{combined data of internal IPD and psuedo IPD from aggregate data.}

\item{trt}{internal trial treatment}

\item{trt_ext}{external trial treatment}

\item{trt_common}{treatment that is shared between internal and external trial. Only applies in anchored comparison.}

\item{break.x.by}{bin parameter for survminer}

\item{endpoint_name}{a character name of the endpoint}
}
\value{
a Kaplan-Meier plot
}
\description{
Function to plot Kaplan-Meier curves using survminer package
}
