% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_km.R
\name{kmplot}
\alias{kmplot}
\title{Kaplan Meier (KM) plot function for anchored and unanchored cases}
\usage{
kmplot(
  ipd_weights,
  tte_dat_ipd,
  ipd_trt_var = "ARM",
  tte_dat_pseudo,
  pseudo_trt_var = "ARM",
  trt_ipd,
  trt_agd,
  trt_common = NULL,
  km_conf_type = "log-log",
  km_layout = c("all", "by_trial", "by_arm"),
  ...
)
}
\arguments{
\item{ipd_weights}{an object returned by \code{estimate_weight}}

\item{tte_dat_ipd}{a data frame of individual patient data (IPD) of internal trial, contain at least "USUBJID", "EVENT", "TIME" columns and a column indicating treatment assignment}

\item{ipd_trt_var}{a string, column name in \code{dat_ipd} that contains the treatment assignment}

\item{tte_dat_pseudo}{a data frame of pseudo IPD by digitized KM curves of external trial (for time-to-event endpoint), contain at least "EVENT", "TIME"}

\item{pseudo_trt_var}{a string, column name in \code{dat_ipd} that contains the treatment assignment}

\item{trt_ipd}{a string, name of the interested investigation arm in internal trial \code{dat_igd} (real IPD)}

\item{trt_agd}{a string, name of the interested investigation arm in external trial \code{dat_pseudo} (pseudo IPD)}

\item{trt_common}{a string, name of the common comparator in internal and external trial, by default is NULL, indicating unanchored case}

\item{km_conf_type}{a string, pass to \code{conf.type} of \code{survfit}}

\item{km_layout}{a string, only applicable for unachored case (\code{trt_common = NULL}), indicated the desired layout of output KM curve.}

\item{...}{other arguments in \code{basic_kmplot}}
}
\value{
In unanchored case, a KM plot with risk set table. In anchored case, depending on \code{km_layout},
\itemize{
\item if "by_trial", 2 by 1 plot, first all KM curves (incl. weighted) in IPD trial, and then KM curves in AgD trial, with risk set table.
\item if "by_arm", 2 by 1 plot, first KM curves of \code{trt_agd} and  \code{trt_ipd} (with and without weights), and then KM cuvers of \code{trt_common} in AgD trial and IPD trial (with and without weights). Risk set table is appended.
\item if "all", 2 by 2 plot, all plots in "by_trial" and "by_arm" without risk set table appended.
}
}
\description{
It is wrapper function of \code{basic_kmplot}.
The argument setting is similar to \code{maic_anchored} and \code{maic_unanchored},
and it is used in those two functions.
}
