% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_km2.R
\name{kmplot2}
\alias{kmplot2}
\title{Kaplan-Meier (KM) plot function for anchored and unanchored cases using ggplot}
\usage{
kmplot2(
  weights_object,
  tte_ipd,
  tte_pseudo_ipd,
  trt_ipd,
  trt_agd,
  trt_common = NULL,
  trt_var_ipd = "ARM",
  trt_var_agd = "ARM",
  km_conf_type = "log-log",
  km_layout = c("all", "by_trial", "by_arm"),
  time_scale,
  ...
)
}
\arguments{
\item{weights_object}{an object returned by \code{estimate_weight}}

\item{tte_ipd}{a data frame of individual patient data (IPD) of internal trial, contain at least \code{"USUBJID"}, \code{"EVENT"}, \code{"TIME"} columns and a column indicating treatment assignment}

\item{tte_pseudo_ipd}{a data frame of pseudo IPD by digitized KM curves of external trial (for time-to-event endpoint), contain at least \code{"EVENT"}, \code{"TIME"}}

\item{trt_ipd}{a string, name of the interested investigation arm in internal trial \code{dat_igd} (real IPD)}

\item{trt_agd}{a string, name of the interested investigation arm in external trial \code{dat_pseudo} (pseudo IPD)}

\item{trt_common}{a string, name of the common comparator in internal and external trial, by default is NULL, indicating unanchored case}

\item{trt_var_ipd}{a string, column name in \code{tte_ipd} that contains the treatment assignment}

\item{trt_var_agd}{a string, column name in \code{tte_pseudo_ipd} that contains the treatment assignment}

\item{km_conf_type}{a string, pass to \code{conf.type} of \code{survfit}}

\item{km_layout}{a string, only applicable for unanchored case (\code{trt_common = NULL}), indicated the desired layout of output KM curve.}

\item{time_scale}{a string, time unit of median survival time, taking a value of 'years', 'months', 'weeks' or 'days'}

\item{...}{other arguments in \code{basic_kmplot2}}
}
\value{
In unanchored case, a KM plot with risk set table. In anchored case, depending on \code{km_layout},
\itemize{
\item if "by_trial", 2 by 1 plot, first all KM curves (incl. weighted) in IPD trial, and then KM curves in AgD trial, with risk set table.
\item if "by_arm", 2 by 1 plot, first KM curves of \code{trt_agd} and  \code{trt_ipd} (with and without weights), and then KM cuvers of \code{trt_common} in AgD trial and IPD trial (with and without weights). Risk set table is appended.
\item if "all", 2 by 2 plot, all plots in "by_trial" and "by_arm" without risk set table appended.
}
}
\description{
This is wrapper function of \code{basic_kmplot2}.
The argument setting is similar to \code{maic_anchored} and \code{maic_unanchored},
and it is used in those two functions.
}
