% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic_anchored.R
\name{maic_anchored}
\alias{maic_anchored}
\title{Anchored MAIC for continuous, binary and time-to-event endpoint}
\usage{
maic_anchored(
  ipd_weights,
  dat_ipd,
  ipd_trt_var = "arm",
  dat_pseudo,
  pseudo_trt_var = "arm",
  trt_ipd,
  trt_agd,
  trt_common,
  endpoint_type = "tte",
  endpoint_name = "Time to Event Endpoint",
  eff_measure = c("HR", "OR", "RR", "RD", "Diff"),
  time_scale = "months",
  km_conf_type = "log-log"
)
}
\arguments{
\item{ipd_weights}{an object returned by \code{estimate_weight}}

\item{dat_ipd}{a data frame that meet format requirements in 'Details', individual patient data (IPD) of internal trial}

\item{ipd_trt_var}{a string, column name in \code{dat_ipd} that contains the treatment assignment}

\item{dat_pseudo}{a data frame, pseudo IPD from digitized KM curve of external trial (for time-to-event endpoint) or from contingency table (for binary endpoint)}

\item{pseudo_trt_var}{a string, column name in \code{dat_ipd} that contains the treatment assignment}

\item{trt_ipd}{a string, name of the interested investigation arm in internal trial \code{dat_igd} (real IPD)}

\item{trt_agd}{a string, name of the interested investigation arm in external trial \code{dat_pseudo} (pseudo IPD)}

\item{trt_common}{a string, name of the common comparator in internal and external trial}

\item{endpoint_type}{a string, one out of the following "binary", "tte" (time to event), "continuous"}

\item{endpoint_name}{a string, name of time to event endpoint, to be show in the last line of title}

\item{eff_measure}{a string, "Diff" for continuous,"RD" (risk difference), "OR" (odds ratio), "RR" (relative risk) for a binary endpoint; "HR" for a time-to-event endpoint. By default is \code{NULL}, "OR" is used for binary case, otherwise "HR" is used.}

\item{time_scale}{a string, time unit of median survival time, taking a value of 'years', 'months', 'weeks' or 'days'}

\item{km_conf_type}{a string, pass to \code{conf.type} of \code{survfit}}
}
\value{
A list, contains 'descriptive' and 'inferential'
}
\description{
This is a wrapper function to provide adjusted effect estimates and relevant statistics in anchored case
(i.e. there is a common comparator arm in the internal and external trial).
}
\details{
Format requirements for input \code{dat_ipd} and \code{dat_pseudo} are to have the following columns
\itemize{
\item USUBJID - charater, unique subject ID
\item ARM - character or factor column
\item EVENT - numeric, 1 for censored/death, 0 for otherwise
\item TIME - numeric column, observation time of the \code{EVENT}; unit in days
}
}
