% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic_unanchored_tte.R
\name{maic_tte_unanchor}
\alias{maic_tte_unanchor}
\title{Unanchored MAIC for time-to-event endpoint}
\usage{
maic_tte_unanchor(
  useWt,
  dat,
  dat_ext,
  trt,
  trt_ext,
  time_scale = "month",
  endpoint_name = "OS",
  transform = "log"
)
}
\arguments{
\item{useWt}{a numeric vector of individual MAIC weights, length should the same as \code{nrow(dat)}}

\item{dat}{a data frame that meet format requirements in 'Details', individual patient data (IPD) of internal trial}

\item{dat_ext}{a data frame, pseudo IPD from digitized KM curve of external trial}

\item{trt}{a character string, name of the interested treatment in internal trial (real IPD)}

\item{trt_ext}{character string, name of the interested comparator in external trial used to subset
\code{dat_ext} (pseudo IPD)}

\item{time_scale}{a character string, 'year', 'month', 'week' or 'day', time unit of median survival time}

\item{endpoint_name}{a character string, name of the endpoint}

\item{transform}{a character string, pass to \code{\link[survival]{cox.zph}}}
}
\value{
A list of KM plot, analysis table, and diagnostic plot
}
\description{
Unanchored MAIC for time-to-event endpoint
}
\details{
Format requirements for input \code{dat} and \code{dat_ext} are to have the following columns
\itemize{
\item treatment - character or factor column
\item status - logical column, TRUE for censored/death, FALSE for otherwise
\item time - numeric column, observation time of the \code{status}; unit in days
}
}
