% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maicplus.R
\docType{package}
\name{maicplus-package}
\alias{maicplus-package}
\title{maicplus: open source R package for MAIC}
\description{
A package for running matching adjusted indirect comparison
}
\details{
This package facilitates performing matching adjusted indirect comparison
(MAIC) analysis where the endpoint of interest is either time-to-event
(e.g. overall survival) or binary (e.g. objective tumor response).
}
