% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helper.R
\name{medSurv_makeup}
\alias{medSurv_makeup}
\title{helper function: makeup to get median survival time from a \code{survival::survfit} object}
\usage{
medSurv_makeup(km_fit, legend = "before matching", time_scale)
}
\arguments{
\item{km_fit}{returned object from \code{survival::survfit}}

\item{legend}{a character string, name used in 'type' column in returned data frame}

\item{time_scale}{a character string, 'year', 'month', 'week' or 'day', time unit of median survival time}
}
\value{
a data frame with a index column 'type', median survival time and confidence interval
}
\description{
extract and display median survival time with confidence interval
}
\examples{
load(system.file("extdata", "combined_data_tte.rda", package = "maicplus", mustWork = TRUE))
kmobj <- survfit(Surv(TIME, EVENT) ~ ARM, combined_data_tte, conf.type = "log-log")
medSurv <- medSurv_makeup(kmobj, legend = "before matching", time_scale = "day")

}
