% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{merge_two_data}
\alias{merge_two_data}
\title{Function to merge IPD data and pseudo comparator IPD}
\usage{
merge_two_data(
  external = NULL,
  internal = NULL,
  internal_time_name = NULL,
  internal_event_name = NULL,
  internal_response_name = NULL
)
}
\arguments{
\item{external}{pseudo comparator IPD. Should be a data frame.
For a time to event outcome, time, status(i.e. event=1), and ARM should be specified
For a response outcome, response and ARM should be specified.}

\item{internal}{internal IPD data that is returned from \code{\link{estimate_weights}}}

\item{internal_time_name}{how the time variable is named in the internal IPD (for time to event outcome)}

\item{internal_event_name}{how the event variable is named in the internal IPD (for time to event outcome)}

\item{internal_response_name}{how the response variable is named in the internal IPD (for binary outcome)}
}
\value{
Merged dataset with time, event, ARM, and weights for time to event data and response, ARM, and weights for binary.
External ARM is assigned to be the reference treatment in the unanchored case. common treatment is assigned to be reference
treatment for the anchored case.
}
\description{
For a survival outcome, we can digitize Kaplan-Meier curves to obtain
pseudo IPD. For a binomial outcome, we can simulate response data based
on the known proportion of responders. For both types of outcomes,
we need to merge the pseudo comparator IPD with the internal IPD data
to run the regression (i.e. cox regression). This function joins merges
the two data into a single data. For the external pseudo IPD, a weight
of 1 is assigned.
}
