% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{plot_weights}
\alias{plot_weights}
\title{Plot MAIC weights in a histogram with key statistics in legend}
\usage{
plot_weights(wt, main_title = "Unscaled Individual Weigths")
}
\arguments{
\item{wt}{a numeric vector of individual MAIC weights (derived use in \code{\link{cal_weights}})}

\item{main_title}{a character string, main title of the plot}
}
\value{
a plot
}
\description{
Generates a plot given the individuals weights with key summary in top right legend that
includes median weight, effective sample size (ESS), reduction percentage (what percent
ESS takes up in the original sample size),
}
