% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{plot_weights2}
\alias{plot_weights2}
\title{Plot MAIC weights in a histogram with key statistics in legend using ggplot}
\usage{
plot_weights2(
  match_res,
  bin_col = "black",
  vline_col = "red",
  bins = 50,
  print_caption = FALSE,
  caption_width = 80
)
}
\arguments{
\item{match_res}{a numeric vector of individual MAIC weights (derived using \code{\link{estimate_weights}})}

\item{bin_col}{a string, color for the bins of histogram}

\item{vline_col}{a string, color for the vertical line in the histogram}

\item{bins}{number of bin parameter to use in the ggplot}

\item{print_caption}{print a footnote message related to ESS from the NICE survey 2021}

\item{caption_width}{width that is passed onto str_wrap function}
}
\value{
a plot of unscaled and scaled weights
}
\description{
Generates a plot given the individuals weights with key summary in top right legend that includes
median weight, effective sample size (ESS), and reduction percentage (what percent ESS is reduced from the
original sample size).
There are two options of weights provided in \code{\link{estimate_weights}}: unscaled or scaled.
Scaled weights are relative to the original unit weights of each individual.
In other words, a scaled weight greater than 1 means that an individual carries more weight in the
re-weighted population than the original data and a scaled weight less than 1 means that an individual carries
less weight in the re-weighted population than the original data.
}
