% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{plot_weights_ggplot}
\alias{plot_weights_ggplot}
\title{Plot MAIC weights in a histogram with key statistics in legend using ggplot}
\usage{
plot_weights_ggplot(
  weighted_data,
  bin_col,
  vline_col,
  main_title,
  print_caption,
  caption_width,
  bins
)
}
\arguments{
\item{weighted_data}{object returned after calculating weights using \link{estimate_weights}}

\item{bin_col}{a string, color for the bins of histogram}

\item{vline_col}{a string, color for the vertical line in the histogram}

\item{main_title}{Name of scaled weights plot and unscaled weights plot, respectively.}

\item{print_caption}{print a footnote message related to ESS from the NICE survey 2021}

\item{caption_width}{width that is passed onto str_wrap function}

\item{bins}{number of bin parameter to use}
}
\value{
a plot of unscaled and scaled weights
}
\description{
Generates a ggplot histogram of weights. Default is to plot both unscaled and scaled weights on a same graph.
}
