% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucher.R
\name{print.bucher}
\alias{print.bucher}
\title{Report-friendly output format for resutl from Bucher's method}
\usage{
\method{print}{bucher}(output, ci.digits = 2, pval.digits = 3)
}
\arguments{
\item{output}{output from \code{\link{bucher}} function}

\item{ci.digits}{an integer, number of decimal places for point estimate and derived confidence limits}

\item{pval.digits}{an integer, number of decimal places to display Z-test p-value}
}
\value{
a character vector of two elements, first element inf format of 'est (ci_l; ci_u)', second element is Z-test p-value, rounded according to \code{pval.digits}
}
\description{
Report-friendly output format for resutl from Bucher's method
}
