% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{print.check.weights}
\alias{print.check.weights}
\title{Print the data frame from \code{\link{check_weights}}}
\usage{
\method{print}{check.weights}(x, ...)
}
\arguments{
\item{x}{Result object created by \code{\link{network.run}} function}

\item{...}{Additional arguments for printing a data frame}
}
\value{
Returns data frame of summary statistics of weighted and unweighted IPD and aggregate data
}
\description{
Print the data frame from \code{\link{check_weights}}. It also gives a warning
message if median is used to match and no IPD equals to reported
aggregate data median.
}
