% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucher.R
\name{print_bucher}
\alias{print_bucher}
\title{Report-friendly output format for resutl from Bucher's method}
\usage{
print_bucher(output, ci_digits = 2, pval_digits = 3)
}
\arguments{
\item{output}{output from \code{\link{bucher}} function}

\item{ci_digits}{an integer, number of decimal places for point estimate and derived confidence limits}

\item{pval_digits}{an integer, number of decimal places to display Z-test p-value}
}
\value{
a character vector of two elements, first element inf format of 'est (ci_l; ci_u)', second element is Z-test p-value, rounded according to \code{pval_digits}
}
\description{
Report-friendly output format for resutl from Bucher's method
}
