% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process.agd}
\alias{process.agd}
\title{Process and Check Aggregated data format}
\usage{
process.agd(raw.agd)
}
\arguments{
\item{raw.agd}{}
}
\description{
Process and Check Aggregated data format
}
