% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process.ipd}
\alias{process.ipd}
\title{Process Individual Patient data to dummize categorical effective modifiers}
\usage{
process.ipd(raw.ipd, dummize.cols, dummize.ref.level)
}
\arguments{
\item{dummize.ref.level}{}
}
\description{
Process Individual Patient data to dummize categorical effective modifiers
}
