% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_agd}
\alias{process_agd}
\title{Process and Check Aggregated data format}
\usage{
process_agd(raw_agd)
}
\arguments{
\item{raw_agd}{}
}
\description{
Process and Check Aggregated data format
}
