% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_ipd}
\alias{process_ipd}
\title{Process Individual Patient data to dummize categorical effective modifiers}
\usage{
process_ipd(raw_ipd, dummize_cols, dummize_ref_level)
}
\arguments{
\item{dummize_ref_level}{}
}
\description{
Process Individual Patient data to dummize categorical effective modifiers
}
