% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pseudo_ipd_sat}
\alias{pseudo_ipd_sat}
\title{Pseudo individual patient survival data from published study}
\format{
A data frame with 300 rows and 3 columns:
\describe{
\item{Time}{}
\item{Event}{}
\item{ARM}{}
}
}
\usage{
pseudo_ipd_sat
}
\description{
Pseudo individual patient survival data from published study
}
\seealso{
Other unanchored datasets: 
\code{\link{adrs_sat}},
\code{\link{adsl_sat}},
\code{\link{adtte_sat}},
\code{\link{agd}},
\code{\link{centered_ipd_sat}}
}
\concept{unanchored datasets}
\keyword{dataset}
