% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pseudo_ipd_twt}
\alias{pseudo_ipd_twt}
\title{Pseudo individual patient survival data from published two arm study}
\format{
A data frame with 800 rows and 3 columns:
\describe{
\item{Time}{}
\item{Event}{}
\item{ARM}{}
}
}
\usage{
pseudo_ipd_twt
}
\description{
Pseudo individual patient survival data from published two arm study
}
\seealso{
Other anchored datasets: 
\code{\link{adsl_twt}},
\code{\link{adtte_twt}},
\code{\link{agd}},
\code{\link{centered_ipd_twt}}
}
\concept{anchored datasets}
\keyword{dataset}
