% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{report_table_binary}
\alias{report_table_binary}
\title{helper function: sort out a nice report table to summarize binary analysis results}
\usage{
report_table_binary(
  binobj,
  weighted_result = NULL,
  eff_measure = c("OR", "RD", "RR"),
  tag = NULL
)
}
\arguments{
\item{binobj}{object from glm()}

\item{weighted_result}{weighted result object}

\item{eff_measure}{a string, binary effect measure, could be "OR", "RR", "RD"}

\item{tag}{a string, by default NULL, if specified, an extra 1st column is created in the output}
}
\value{
a data frame with sample size, incidence rate, estimate of binary effect measure with
95\% CI and Wald test of hazard ratio
}
\description{
helper function: sort out a nice report table to summarize binary analysis results
}
\examples{
data(adrs_sat)
testdat <- data.frame(Yes = 280, No = 120)
rownames(testdat) <- "B"
pseudo_ipd_binary_sat <- get_pseudo_ipd_binary(
  binary_agd = testdat,
  format = "unstacked"
)
combined_data <- rbind(adrs_sat[, c("USUBJID", "RESPONSE", "ARM")], pseudo_ipd_binary_sat)
combined_data$ARM <- as.factor(combined_data$ARM)

binobj_dat <- glm(RESPONSE ~ ARM, combined_data, family = binomial(link = "logit"))
report_table_binary(binobj_dat, eff_measure = "OR")
}
