% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{report_table_tte}
\alias{report_table_tte}
\title{helper function: sort out a nice report table to summarize survival analysis results}
\usage{
report_table_tte(coxobj, medSurvobj, tag = NULL)
}
\arguments{
\item{coxobj}{returned object from \code{\link[survival]{coxph}}}

\item{medSurvobj}{returned object from \code{\link{medSurv_makeup}}}

\item{tag}{a string, by default NULL, if specified, an extra 1st column is created in the output}
}
\value{
a data frame with sample size, incidence rate, median survival time with 95\% CI, hazard ratio estimate with
95\% CI and Wald test of hazard ratio
}
\description{
helper function: sort out a nice report table to summarize survival analysis results
}
\examples{
data(adtte_sat)
data(pseudo_ipd_sat)
combined_data <- rbind(adtte_sat[, c("TIME", "EVENT", "ARM")], pseudo_ipd_sat)
unweighted_cox <- coxph(Surv(TIME, EVENT == 1) ~ ARM, data = combined_data)

# Derive median survival time
kmobj <- survfit(Surv(TIME, EVENT) ~ ARM, combined_data, conf.type = "log-log")
medSurv <- medSurv_makeup(kmobj, legend = "before matching", time_scale = "day")

report_table_tte(unweighted_cox, medSurv)
}
