% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helper.R
\name{resid_plot}
\alias{resid_plot}
\title{Plot schoenfeld residual plot for a Cox model fit}
\usage{
resid_plot(
  coxobj,
  time_scale = "month",
  log_time = TRUE,
  endpoint_name = "",
  subtitle = ""
)
}
\arguments{
\item{coxobj}{object returned from \code{\link[survival]{coxph}}}

\item{time_scale}{a character string, 'year', 'month', 'week' or 'day', time unit of median survival time}

\item{log_time}{logical, TRUE (default) or FALSE}

\item{endpoint_name}{a character string, name of the endpoint}

\item{subtitle}{a character string, subtitle of the plot}
}
\value{
a plot
}
\description{
Plot schoenfeld residual plot for a Cox model fit
}
