% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helper.R
\name{survfit_makeup}
\alias{survfit_makeup}
\title{Helper function to select a set of variables used for Kaplan-Meier plot}
\usage{
survfit_makeup(km_fit)
}
\arguments{
\item{km_fit}{returned object from \code{survival::survfit}}
}
\value{
a list of data frames of variables from \code{survival::survfit}. Data frames are divided by treatment.
}
\description{
Helper function to select a set of variables used for Kaplan-Meier plot
}
