% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unanchored-data}
\alias{unanchored-data}
\alias{adsl_sat}
\alias{adtte_sat}
\alias{pseudo_ipd_sat}
\alias{agd}
\alias{centered_ipd_sat}
\alias{adrs_sat}
\title{Dummy Single-arm Trial Data for Unanchored MAIC}
\format{
\strong{adsl_sat}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

\strong{adtte_sat}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

\strong{pseudo_ipd_sat}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

\strong{agd}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

\strong{centered_ipd_sat}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

\strong{adrs_sat}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

An object of class \code{data.frame} with 500 rows and 10 columns.

An object of class \code{data.frame} with 300 rows and 3 columns.

An object of class \code{data.frame} with 3 rows and 9 columns.

An object of class \code{data.frame} with 500 rows and 15 columns.

An object of class \code{data.frame} with 500 rows and 5 columns.
}
\source{
These datasets are generated by the dummy_unanchored.R script.
}
\usage{
adsl_sat

adtte_sat

pseudo_ipd_sat

agd

centered_ipd_sat

adrs_sat
}
\description{
These datasets contains adsl_sat, adtte_sat, adrs_sat,
agd (AgD of effect modifiers), pseudo_ipd_sat (AgD, tte data)
}
\keyword{dataset}
\keyword{datasets}
