% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weighted_sat}
\alias{weighted_sat}
\title{Weighted object for single arm trial data}
\format{
A \code{maicplus_estimate_weights} object created by \code{\link[=estimate_weights]{estimate_weights()}} containing
\describe{
\item{data}{patient level data with weights}
\item{centered_colnames}{Columns used in MAIC}
\item{nr_missing}{Number of observations with missing data}
\item{ess}{Expected sample size}
\item{opt}{Information from \code{optim} from weight calculation}
\item{boot}{Parameters and bootstrap sample weights, \code{NULL} in this object}
}
}
\usage{
weighted_sat
}
\description{
Weighted object for single arm trial data
}
\seealso{
Other unanchored datasets: 
\code{\link{adrs_sat}},
\code{\link{adsl_sat}},
\code{\link{adtte_sat}},
\code{\link{agd}},
\code{\link{centered_ipd_sat}},
\code{\link{pseudo_ipd_sat}}
}
\concept{unanchored datasets}
\keyword{dataset}
