% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weighted_twt}
\alias{weighted_twt}
\title{Weighted object for two arm trial data}
\format{
A \code{maicplus_estimate_weights} object created by \code{\link[=estimate_weights]{estimate_weights()}} containing
\describe{
\item{data}{patient level data with weights}
\item{centered_colnames}{Columns used in MAIC}
\item{nr_missing}{Number of observations with missing data}
\item{ess}{Expected sample size}
\item{opt}{Information from \code{optim} from weight calculation}
\item{boot}{Parameters and bootstrap sample weights, \code{NULL} in this object}
}
}
\usage{
weighted_twt
}
\description{
Weighted object for two arm trial data
}
\seealso{
Other anchored datasets: 
\code{\link{adsl_twt}},
\code{\link{adtte_twt}},
\code{\link{agd}},
\code{\link{centered_ipd_twt}},
\code{\link{pseudo_ipd_twt}}
}
\concept{anchored datasets}
\keyword{dataset}
