% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adrs_twt}
\alias{adrs_twt}
\title{Binary outcome data from two arm trial}
\format{
A data frame with 1000 rows and 5 columns:
\describe{
\item{USUBJID}{Unique subject identifiers for patients.}
\item{ARM}{Assigned treatment arm, \code{"A"}, \code{"C"}.}
\item{AVAL}{Analysis value, in this dataset an indicator of response.}
\item{PARAM}{Parameter type of \code{AVAL}.}
\item{RESPONSE}{Indicator of response.}
}
}
\usage{
adrs_twt
}
\description{
Binary outcome data from two arm trial
}
\seealso{
Other anchored datasets: 
\code{\link{adsl_twt}},
\code{\link{adtte_twt}},
\code{\link{agd}},
\code{\link{centered_ipd_twt}},
\code{\link{pseudo_ipd_twt}},
\code{\link{weighted_twt}}
}
\concept{anchored datasets}
\keyword{dataset}
