% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl_sat}
\alias{adsl_sat}
\title{Patient data from single arm study}
\format{
a data frame with 500 rows and 8 columns:
\describe{
\item{USUBJID}{}
\item{ARM}{}
\item{AGE}{}
\item{SEX}{}
\item{SMOKE}{}
\item{ECOG0}{}
\item{N_PR_THER}{}
\item{SEX_MALE}{}
}
}
\usage{
adsl_sat
}
\description{
Patient data from single arm study
}
\seealso{
Other unanchored datasets: 
\code{\link{adrs_sat}},
\code{\link{adtte_sat}},
\code{\link{agd}},
\code{\link{centered_ipd_sat}},
\code{\link{pseudo_ipd_sat}}
}
\concept{unanchored datasets}
\keyword{dataset}
