% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adtte_sat}
\alias{adtte_sat}
\title{Survival data from single arm trial}
\format{
A data frame with 500 rows and 10 columns:
\describe{
\item{USUBJID}{Unique subject identifiers for patients.}
\item{ARM}{Assigned treatment arm, \code{"A"}.}
\item{AVAL}{Analysis value which in this dataset overall survival time in days.}
\item{AVALU}{Unit of \code{AVAL}.}
\item{PARAMCD}{Paramater code of \code{AVAL}, \code{"OS"}.}
\item{PARAM}{Parameter name of \code{AVAL}, \verb{"Overall Survival}.}
\item{CNSR}{Censoring indicator \code{0}/\code{1}.}
\item{TIME}{Survival time in days.}
\item{EVENT}{Event indicator \code{0}/\code{1}.}
}
}
\usage{
adtte_sat
}
\description{
Survival data from single arm trial
}
\seealso{
Other unanchored datasets: 
\code{\link{adrs_sat}},
\code{\link{adsl_sat}},
\code{\link{agd}},
\code{\link{centered_ipd_sat}},
\code{\link{pseudo_ipd_sat}},
\code{\link{weighted_sat}}
}
\concept{unanchored datasets}
\keyword{dataset}
