% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adtte_twt}
\alias{adtte_twt}
\title{Survival data from two arm trial}
\format{
A data frame with 1000 rows and 10 columns:
\describe{
\item{USUBJID}{Unique subject identifiers for patients.}
\item{ARM}{Assigned treatment arm, \code{"A"}, \code{"C"}.}
\item{AVAL}{Analysis value which in this dataset overall survival time in days.}
\item{AVALU}{Unit of \code{AVAL}.}
\item{PARAMCD}{Parameter code of \code{AVAL}, \code{"OS"}.}
\item{PARAM}{Parameter name of \code{AVAL}, \verb{"Overall Survival}.}
\item{CNSR}{Censoring indicator \code{0}/\code{1}.}
\item{TIME}{Survival time in days.}
\item{EVENT}{Event indicator \code{0}/\code{1}.}
}
}
\usage{
adtte_twt
}
\description{
Survival data from two arm trial
}
\seealso{
Other anchored datasets: 
\code{\link{adrs_twt}},
\code{\link{adsl_twt}},
\code{\link{agd}},
\code{\link{centered_ipd_twt}},
\code{\link{pseudo_ipd_twt}},
\code{\link{weighted_twt}}
}
\concept{anchored datasets}
\keyword{dataset}
