% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agd}
\alias{agd}
\title{Aggregate effect modifier data from published study}
\format{
A data frame with 3 rows and 9 columns:
\describe{
\item{STUDY}{The study name, Study_XXXX}
\item{ARM}{Study arm name or total}
\item{N}{Number of observations in study arm}
\item{AGE_MEAN}{Mean age in study arm}
\item{AGE_MEDIAN}{Median age in study arm}
\item{AGE_SD}{Standard deviation of age in study arm}
\item{SEX_MALE_COUNT}{Number of male patients}
\item{ECOG0_COUNT}{Number of patients with ECOG score = 0}
\item{SMOKE_COUNT}{Number of smokers}
\item{N_PR_THER_MEDIAN}{Median number of prior therapies}
}
}
\usage{
agd
}
\description{
This data is formatted to be used in \code{\link[=center_ipd]{center_ipd()}}.
}
\seealso{
Other unanchored datasets: 
\code{\link{adrs_sat}},
\code{\link{adsl_sat}},
\code{\link{adtte_sat}},
\code{\link{centered_ipd_sat}},
\code{\link{pseudo_ipd_sat}},
\code{\link{weighted_sat}}

Other anchored datasets: 
\code{\link{adrs_twt}},
\code{\link{adsl_twt}},
\code{\link{adtte_twt}},
\code{\link{centered_ipd_twt}},
\code{\link{pseudo_ipd_twt}},
\code{\link{weighted_twt}}
}
\concept{anchored datasets}
\concept{unanchored datasets}
\keyword{dataset}
