% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anchored-data}
\alias{anchored-data}
\alias{adsl_twt}
\alias{adtte_twt}
\alias{pseudo_ipd_twt}
\alias{centered_ipd_twt}
\title{Dummy Trial Data for anchored MAIC}
\format{
\strong{adsl_twt}: a data frame with 1000 rows and 9 columns:
\describe{
\item{X}{The ID of the patients.}
\item{USUBJID}{Unique subject identifiers for patients.}
\item{ARM}{Assigned treatment arm.}
\item{AGE}{Age in years at baseline.}
\item{SEX}{Sex of patient recorded as character "Male"/"Female"}
\item{SMOKE}{Smoking status at baseline as integer 0/1}
\item{ECOG0}{ECOG score at baseline as integer 0/1}
\item{N_PR_THER}{Number of prior therapies received as integer 1, 2, 3, 4}
\item{SEX_MALE}{Indicator of SEX == "Male" as numeric 1/0}
}

\strong{adtte_twt}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

\strong{pseudo_ipd_twt}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}

\strong{centered_ipd_twt}: a data frame with xx rows and x columns:
\describe{
\item{id}{The ID of the patients.}
}
}
\source{
These datasets are generated by the dummy_anchored.R script.
}
\usage{
adsl_twt

adtte_twt

pseudo_ipd_twt

centered_ipd_twt
}
\description{
These datasets contains adsl_twt, adtte_twt,
centered_ipd_twt,  pseudo_ipd_twt (AgD, tte data)
}
\keyword{dataset}
\keyword{datasets}
