% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_HR}
\alias{bootstrap_HR}
\title{Bootstrapping weighted hazard ratios to obtain confidence interval}
\usage{
bootstrap_HR(
  ipd_centered,
  i,
  centered_colnames,
  pseudo_ipd = NULL,
  model,
  min_weight = 1e-04,
  ref_treat = NULL
)
}
\arguments{
\item{ipd_centered}{A data frame containing individual patient data
from the internal IPD study. This data frame should have already been
centered using \link{center_ipd} function}

\item{i}{Index used to select a sample within \code{\link{boot}}.}

\item{centered_colnames}{A character vector giving the names of the covariates to use in matching. These names must match the column names in internal dataset.}

\item{model}{Survival model to run to calculate HR with correct time, event, and #'   ARM names}

\item{min_weight}{A numeric value that defines the minimum weight allowed.
This value (default 0.0001) will replace weights estimated at 0 in a sample.}

\item{ref_treat}{Reference treatment for the survival model}

\item{psuedo_ipd}{A data frame containing pseudo individual patient data
from the comparator study needed to derive the relative treatment effect.
Pseudo IPD is only used for unanchored case. pseudo_ipd should have time, event,
and ARM variable with a same name and order as that of ipd_centered}
}
\value{
The HR bootstraps
}
\description{
This function uses \code{\link{boot}} to resample internal IPD data
to obtain bootstrapped hazard ratio values which is used to form confidence
interval. Note that due to lack of full IPD of covariates in the
aggregate data, pseudo IPD will not be bootstrapped.
}
\examples{
# Unanchored example
load(system.file("extdata", "ipd.rda", package = "maicplus", mustWork = TRUE))
load(system.file("extdata", "agd.rda", package = "maicplus", mustWork = TRUE))
ipd_centered <- center_ipd(ipd = adsl, agd = agd)

pseudo_ipd <- read.csv(system.file("extdata", "psuedo_IPD.csv", package = "maicplus", mustWork = TRUE))
pseudo_ipd$ARM <- "B" # Need to specify ARM for pseudo ipd

# Need to ensure pseudo_ipd and ipd_matched have same names
colnames(pseudo_ipd) <- c("TIME", "EVENT", "ARM")
\dontrun{
HR_bootstraps <- boot(
  data = ipd_centered,
  statistic = bootstrap_HR,
  centered_colnames = centered_colnames,
  pseudo_ipd = pseudo_ipd,
  model = Surv(TIME, EVENT == 1) ~ ARM,
  ref_treat = "B",
  R = 1000
)
}
}
