% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_HR}
\alias{bootstrap_HR}
\title{Bootstrapping weighted hazard ratios to obtain confidence interval}
\usage{
bootstrap_HR(
  ipd_centered,
  i,
  centered_colnames,
  internal_time_name = "TIME",
  internal_event_name = "EVENT",
  pseudo_ipd = NULL,
  min_weight = 1e-04
)
}
\arguments{
\item{ipd_centered}{A data frame containing individual patient data
from the internal IPD study. This data frame should have already been
centered using \link{center_ipd} function}

\item{i}{Index used to select a sample within \code{\link{boot}}.}

\item{centered_colnames}{A character vector giving the names of the covariates to use
in matching. These names must match the column names in internal dataset.}

\item{internal_time_name}{name of the time variable in ipd_centered (for time to event outcome)}

\item{internal_event_name}{name of the event variable in ipd_centered (for time to event outcome)}

\item{min_weight}{A numeric value that defines the minimum weight allowed.
This value (default 0.0001) will replace weights estimated at 0 in a sample.}

\item{psuedo_ipd}{A data frame containing pseudo individual patient data
from the comparator study needed to derive the relative treatment effect.
Pseudo IPD is only used for unanchored case.}
}
\value{
The HR bootstraps
}
\description{
This function uses \code{\link{boot}} to resample internal IPD data
to obtain bootstrapped hazard ratio values which is used to form confidence
interval. Note that due to lack of full IPD with covariates in the
aggregate data, pseudo IPD will not be bootstrapped.
}
\examples{
# Unanchored example
load(system.file("extdata","ipd_centered.rda", package = "maicplus", mustWork = TRUE))
centered_colnames <- c("AGE", "AGE_SQUARED", "SEX_MALE", "ECOG0", "SMOKE", "N_PR_THER_MEDIAN")
centered_colnames <- paste0(centered_colnames, "_CENTERED")
pseudo_ipd <- read.csv(system.file("extdata", "psuedo_IPD.csv", package = "maicplus", mustWork = TRUE))
pseudo_ipd$ARM <- "B" #Need to specify ARM for pseudo ipd

HR_bootstraps <- boot(data = ipd_centered,
  statistic = bootstrap_HR, 
  centered_colnames = centered_colnames, 
  internal_time_name = "TIME", 
  internal_event_name = "EVENT",
  pseudo_ipd = pseudo_ipd,
  R = 1000)

}
