% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_HR}
\alias{bootstrap_HR}
\title{Bootstrapping for MAIC weighted hazard ratios}
\usage{
bootstrap_HR(
  ipd_centered,
  i,
  centered_colnames,
  internal_time_name = "TIME",
  internal_event_name = "EVENT",
  external,
  min_weight = 1e-04
)
}
\arguments{
\item{ipd_centered}{A data frame containing individual patient data
from the internal IPD study. This data frame should have already been
centered using \link{center_ipd} function}

\item{i}{Index used to select a sample within \code{\link{boot}}.}

\item{centered_colnames}{A character vector giving the names of the covariates to use
in matching. These names must match the column names in internal dataset.}

\item{internal_time_name}{name of the time variable in ipd_centered (for time to event outcome)}

\item{internal_event_name}{name of the event variable in ipd_centered (for time to event outcome)}

\item{min_weight}{A numeric value that defines the minimum weight allowed.
This value (default 0.0001) will replace weights estimated at 0 in a sample.}

\item{psuedo_ipd}{A data frame containing pseudo individual patient data
from the comparator study needed to derive the relative treatment effect.}
}
\value{
The HR as a numeric value.
}
\description{
Bootstrapping for MAIC weighted hazard ratios
}
\details{
This function is intended to be used in conjunction with the
\code{\link{boot}} function to return the statistic to be
bootstrapped. In this case by performing MAIC weighting using
{\link{estimate_weights}} and returning a weighted hazard ratio (HR) from a
Cox proportional hazards model. This is used as the 'statistic' argument in
the boot function.
}
