% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucher.R
\name{bucher}
\alias{bucher}
\title{Bucher method for adjusted treatment effect}
\usage{
bucher(trt, com, conf_lv = 0.95)
}
\arguments{
\item{trt}{a list with two named scalars for the study with interested experimental arm,
one named \code{'est'} for the point estimate, and the other named \code{'se'} for the standard error}

\item{com}{same as \code{trt}, but for the study with interested control arm}

\item{conf_lv}{a numerical scalar, prescribe confidence level to derive two-sided
confidence interval for the adjusted treatment effect}
}
\value{
a list with 5 elements,
\describe{
\item{est}{a scalar, point estimate of the adjusted treatment effect}
\item{se}{a scalar, standard error of the adjusted treatment effect (i.e. \code{est} in return)}
\item{ci_l}{a scalar, lower confidence limit of a two-sided CI with prescribed nominal level by \code{conf_lv}}
\item{ci_u}{a scalar, upper confidence limit of a two-sided CI with prescribed nominal level by \code{conf_lv}}
\item{pval}{p-value of Z-test, with null hypothesis that \code{est} is zero}
}
}
\description{
Given two estimated treatment effects of A vs. C and B vs. C
(i.e. two point estimates and corresponding standard errors),
derive the adjusted treatment of A vs. B using Bucher method,
with two-sided confidence limits and Z-test p-value
}
\examples{
trt <- list(est = 1.1, se = 0.2)
com <- list(est = 1.3, se = 0.18)
bucher(trt, com, conf_lv = 0.9)
}
