% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centered_ipd_sat}
\alias{centered_ipd_sat}
\title{Centered patient data from single arm trial}
\format{
A data frame with 500 rows and 14 columns:
\describe{
\item{USUBJID}{}
\item{ARM}{}
\item{AGE}{}
\item{SEX}{}
\item{SMOKE}{}
\item{ECOG0}{}
\item{N_PR_THER}{}
\item{SEX_MALE}{}
\item{AGE_CENTERED}{}
\item{AGE_MEDIAN_CENTERED}{}
\item{AGE_SQUARED_CENTERED}{}
\item{SEX_MALE_CENTERED}{}
\item{ECOG0_CENTERED}{}
\item{SMOKE_CENTERED}{}
}
}
\usage{
centered_ipd_sat
}
\description{
Centered patient data from single arm trial
}
\seealso{
Other unanchored datasets: 
\code{\link{adrs_sat}},
\code{\link{adsl_sat}},
\code{\link{adtte_sat}},
\code{\link{agd}},
\code{\link{pseudo_ipd_sat}}
}
\concept{unanchored datasets}
\keyword{dataset}
