% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centered_ipd_twt}
\alias{centered_ipd_twt}
\title{Centered patient data from two arm trial}
\format{
A data frame with 1000 rows and 15 columns:
\describe{
\item{X}{}
\item{USUBJID}{}
\item{ARM}{}
\item{AGE}{}
\item{SEX}{}
\item{SMOKE}{}
\item{ECOG0}{}
\item{N_PR_THER}{}
\item{SEX_MALE}{}
\item{AGE_CENTERED}{}
\item{AGE_MEDIAN_CENTERED}{}
\item{AGE_SQUARED_CENTERED}{}
\item{SEX_MALE_CENTERED}{}
\item{ECOG0_CENTERED}{}
\item{SMOKE_CENTERED}{}
}
}
\usage{
centered_ipd_twt
}
\description{
Centered patient data from two arm trial
}
\seealso{
Other anchored datasets: 
\code{\link{adsl_twt}},
\code{\link{adtte_twt}},
\code{\link{agd}},
\code{\link{pseudo_ipd_twt}}
}
\concept{anchored datasets}
\keyword{dataset}
