% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{check_weights}
\alias{check_weights}
\alias{print.maicplus_check_weights}
\title{Check to see if weights are optimized correctly}
\usage{
check_weights(weighted_data, processed_agd)

\method{print}{maicplus_check_weights}(
  x,
  mean_digits = 2,
  prop_digits = 2,
  sd_digits = 3,
  digits = getOption("digits"),
  ...
)
}
\arguments{
\item{processed_agd}{a data frame, object returned after using \code{\link{process_agd}} or
aggregated data following the same naming convention}

\item{x}{object from \link{check_weights}}

\item{mean_digits}{number of digits for rounding mean columns in the output}

\item{prop_digits}{number of digits for rounding proportion columns in the output}

\item{sd_digits}{number of digits for rounding mean columns in the output}

\item{digits}{minimal number of significant digits, see \link{print.default}.}

\item{...}{further arguments to \link{print.data.frame}}

\item{match_res}{object returned after calculating weights using \code{\link{estimate_weights}}}
}
\value{
data.frame of weighted and unweighted covariate averages of the IPD, and average of aggregate data
}
\description{
This function checks to see if the optimization is done properly by checking the covariate averages
before and after adjustment.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(maicplus_check_weights)}: Print method for check_weights objects

}}
\examples{
load(system.file("extdata", "agd.rda", package = "maicplus", mustWork = TRUE))
load(system.file("extdata", "match_res.rda", package = "maicplus", mustWork = TRUE))
outdata <- check_weights(match_res, processed_agd = agd)
print(outdata)

}
