% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{check_weights}
\alias{check_weights}
\title{Check to see if weights are optimized correctly}
\usage{
check_weights(optimized, match_cov, digits = 2)
}
\arguments{
\item{optimized}{object returned after calculating weights using \code{\link{cal_weights}}}

\item{match_cov}{covariates that should be checked to see if the IPD weighted average matches the aggregate data
average. This could be same set of variables that were used to match or it can include variables that were not
included to match (i.e. stratification variables)}

\item{digits}{number of digits for rounding summary table}
}
\value{
data.frame of weighted and unweighted covariate averages of the IPD
}
\description{
This function checks to see if the optimization is done properly by checking the covariate averages
before and after adjustment.
}
