% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{complete_agd}
\alias{complete_agd}
\title{Calculate pooled arm statistics in Aggregated Data (AgD) based on arm-specific statistics}
\usage{
complete_agd(use_agd)
}
\arguments{
\item{use_agd}{}
}
\description{
Calculate pooled arm statistics in Aggregated Data (AgD) based on arm-specific statistics
}
