% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{ext_tte_transfer}
\alias{ext_tte_transfer}
\title{helper function: transform TTE ADaM data to suitable input for survival R package}
\usage{
ext_tte_transfer(dd, time_scale = "months", trt = NULL)
}
\arguments{
\item{dd}{data frame, ADTTE read via \code{haven::read_sas}}

\item{time_scale}{a character string, 'years', 'months', 'weeks' or 'days', time unit of median survival time}

\item{trt}{values to include in treatment column}
}
\value{
a data frame that can be used as input to \code{survival::Surv}
}
\description{
helper function: transform TTE ADaM data to suitable input for survival R package
}
\keyword{internal}
