% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helper.R
\name{km_plot}
\alias{km_plot}
\title{helper function: KM plot with unadjusted and adjusted KM}
\usage{
km_plot(
  km_fit_before,
  km_fit_after = NULL,
  time_scale,
  trt,
  trt_ext,
  endpoint_name = "",
  line_col = c("#5450E4", "#00857C", "#6ECEB2")
)
}
\arguments{
\item{km_fit_before}{returned object from \code{survival::survfit} before adjustment}

\item{km_fit_after}{returned object from \code{survival::survfit} after adjustment}

\item{time_scale}{time unit of median survival time, taking a value of 'year', 'month', 'week' or 'day'}

\item{trt}{internal trial treatment}

\item{trt_ext}{external trial treatment}

\item{endpoint_name}{name of the endpoint}

\item{line_col}{color of the line curves with the order of external, internal unadjusted, and internal adjusted}
}
\value{
a Kaplan-Meier plot
}
\description{
helper function: KM plot with unadjusted and adjusted KM
}
\examples{
load(system.file("extdata", "combined_data_tte.rda", package = "maicplus", mustWork = TRUE))
kmobj <- survfit(Surv(TIME, EVENT) ~ ARM, combined_data_tte, conf.type = "log-log")
kmobj_adj <- survfit(Surv(TIME, EVENT) ~ ARM, combined_data_tte, weights = combined_data_tte$weights, conf.type = "log-log")
par(cex.main=0.85)
km_plot(kmobj, kmobj_adj, time_scale = "month", trt = "A", trt_ext = "B", endpoint_name = "OS")
}
