% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helper.R
\name{km_plot}
\alias{km_plot}
\title{helper function: KM plot with unadjusted and adjusted KM}
\usage{
km_plot(
  km_fit_before,
  km_fit_after = NULL,
  time_scale,
  trt,
  trt_ext,
  endpoint_name = ""
)
}
\arguments{
\item{km_fit_before}{returned object from \code{survival::survfit} before adjustment}

\item{km_fit_after}{returned object from \code{survival::survfit} after adjustment}

\item{time_scale}{a character string, 'year', 'month', 'week' or 'day', time unit of median survival time}

\item{trt}{a character string, name of the interested treatment in internal trial (real IPD)}

\item{trt_ext}{character string, name of the interested comparator in external trial used to subset \code{dat_ext} (pseudo IPD)}

\item{endpoint_name}{a character string, name of the endpoint}
}
\value{
a KM plot
}
\description{
helper function: KM plot with unadjusted and adjusted KM
}
