% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helper.R
\name{km_plot2}
\alias{km_plot2}
\title{Function to plot Kaplan-Meier curves using survminer package}
\usage{
km_plot2(combined_data, trt, trt_ext, trt_common = NULL, endpoint_name = "")
}
\arguments{
\item{combined_data}{combined data}

\item{trt}{a character string, name of the interested treatment in internal trial (real IPD)}

\item{trt_ext}{character string, name of the interested comparator in external trial used to subset \code{dat_ext} (pseudo IPD)}

\item{endpoint_name}{a character string, name of the endpoint}

\item{time_scale}{a character string, 'year', 'month', 'week' or 'day', time unit of median survival time}
}
\value{
a KM plot
}
\description{
Function to plot Kaplan-Meier curves using survminer package
}
