% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ph_diagnosis.R
\name{log_cum_haz_plot}
\alias{log_cum_haz_plot}
\title{Plot of Log Cumulative Hazard Rate}
\usage{
log_cum_haz_plot(
  km_fit,
  time_scale,
  log_time = TRUE,
  endpoint_name = "",
  subtitle = "",
  exclude_censor = TRUE
)
}
\arguments{
\item{km_fit}{returned object from \code{survival::survfit}}

\item{time_scale}{a character string, 'year', 'month', 'week' or 'day', time unit of median survival time}

\item{log_time}{logical, TRUE (default) or FALSE}

\item{endpoint_name}{a character string, name of the endpoint}

\item{subtitle}{a character string, subtitle of the plot}

\item{exclude_censor}{logical, should censored data point be plotted}
}
\value{
a plot of log cumulative hazard rate
}
\description{
a diagnosis plot for proportional hazard assumption, versus log-time (default) or time
}
\examples{
library(survival)
load(system.file("extdata", "combined_data_tte.rda", package = "maicplus", mustWork = TRUE))
kmobj <- survfit(Surv(TIME, EVENT) ~ ARM, combined_data_tte, conf.type = "log-log")
log_cum_haz_plot(kmobj,
  time_scale = "month", log_time = TRUE,
  endpoint_name = "OS", subtitle = "(Before Matching)"
)
}
