% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic_anchored.R
\name{maic_anchored}
\alias{maic_anchored}
\title{Anchored MAIC for continuous, binary and time-to-event endpoint}
\usage{
maic_anchored(
  useWt,
  dat_ipd,
  ipd_trt_var = "arm",
  dat_pseudo,
  pseudo_trt_var = "arm",
  trt_ipd,
  trt_agd,
  trt_common,
  endpoint_type = "tte",
  endpoint_name = "OS",
  eff_measure = c("HR", "OR", "RR", "RD", "Diff"),
  full_output = TRUE,
  time_scale = "month",
  km_plot_type = c("basic", "ggplot"),
  km_conf_type = "log-log",
  km_layout = c("all", "by_trial", "by_arm"),
  transform = "log"
)
}
\arguments{
\item{useWt}{a numeric vector of individual MAIC weights, length should the same as \code{nrow(dat_igd)}}

\item{dat_ipd}{a data frame that meet format requirements in 'Details', individual patient data (IPD) of internal trial}

\item{ipd_trt_var}{a string, column name in \code{dat_ipd} that contains the treatment assignment}

\item{dat_pseudo}{a data frame, pseudo IPD from digitized KM curve of external trial (for time-to-event endpoint) or from contingency table (for binary endpoint)}

\item{pseudo_trt_var}{a string, column name in \code{dat_ipd} that contains the treatment assignment}

\item{trt_ipd}{a string, name of the interested investigational arm in internal trial \code{dat_igd} (real IPD)}

\item{trt_agd}{a string, name of the interested investigational arm in external trial \code{dat_pseudo} (pseudo IPD)}

\item{trt_common}{a string, name of the common comparator in internal and external trial}

\item{endpoint_type}{a string, one out of the following "binary", "tte" (time to event), "continuous"}

\item{endpoint_name}{a string, name of the endpoint, used in the reported tables and graphs}

\item{eff_measure}{a string, "Diff" for continuous,"RD" (risk difference), "OR" (odds ratio), "RR" (relative risk) for a binary endpoint; "HR" for a time-to-event endpoint. By default is \code{NULL}, "OR" is used for binary case, otherwise "HR" is used.}

\item{full_output}{logical, if FALSE, KM plot and PH diagnosis (for time-to-event endpoint) will be muted. The function will return only a simplified table of adjusted effect estimates; only relevant when \code{endpoint_type}='tte'}

\item{time_scale}{a string, 'year', 'month', 'week' or 'day', time unit of median survival time; only relevant when \code{binary}=FALSE}

\item{transform}{a string, pass to \code{\link[survival]{cox.zph}}; only relevant when \code{endpoint_type}='tte'}
}
\value{
A list of KM plot, analysis table, and diagnostic plot
}
\description{
This is a wrapper function to provide adjusted effect estimates and relevant statistics in anchored case
(i.e. there is a common comparator arm in the internal and external trial).
}
\details{
Format requirements for input \code{dat_ipd} and \code{dat_pseudo} are to have the following columns
\itemize{
\item arm - character or factor column
\item status - logical column, TRUE for censored/death, FALSE for otherwise
\item time - numeric column, observation time of the \code{status}; unit in days
}
}
