% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic_anchored.R
\name{maic_anchored}
\alias{maic_anchored}
\title{Anchored MAIC for binary and time-to-event endpoint}
\usage{
maic_anchored(
  weights_object,
  ipd,
  pseudo_ipd,
  trt_ipd,
  trt_agd,
  trt_common,
  trt_var_ipd = "ARM",
  trt_var_agd = "ARM",
  endpoint_type = "tte",
  endpoint_name = "Time to Event Endpoint",
  eff_measure = c("HR", "OR", "RR", "RD", "Diff"),
  time_scale = "months",
  km_conf_type = "log-log",
  boot_ci_type = c("norm", "basic", "stud", "perc", "bca")
)
}
\arguments{
\item{weights_object}{an object returned by \code{estimate_weight}}

\item{ipd}{a data frame that meet format requirements in 'Details', individual patient data (IPD) of internal trial}

\item{pseudo_ipd}{a data frame, pseudo IPD from digitized KM curve of external trial (for time-to-event endpoint) or
from contingency table (for binary endpoint)}

\item{trt_ipd}{a string, name of the interested investigation arm in internal trial \code{ipd} (internal IPD)}

\item{trt_agd}{a string, name of the interested investigation arm in external trial \code{pseudo_ipd} (pseudo IPD)}

\item{trt_common}{a string, name of the common comparator in internal and external trial}

\item{trt_var_ipd}{a string, column name in \code{ipd} that contains the treatment assignment}

\item{trt_var_agd}{a string, column name in \code{ipd} that contains the treatment assignment}

\item{endpoint_type}{a string, one out of the following "binary", "tte" (time to event)}

\item{endpoint_name}{a string, name of time to event endpoint, to be show in the last line of title}

\item{eff_measure}{a string, "RD" (risk difference), "OR" (odds ratio), "RR" (relative risk) for a binary endpoint;
"HR" for a time-to-event endpoint. By default is \code{NULL}, "OR" is used for binary case, otherwise "HR" is used.}

\item{time_scale}{a string, time unit of median survival time, taking a value of 'years', 'months', 'weeks' or
'days'. NOTE: it is assumed that values in TIME column of \code{ipd} and \code{pseudo_ipd} is in the unit of days}

\item{km_conf_type}{a string, pass to \code{conf.type} of \code{survfit}}

\item{boot_ci_type}{a string, one of \code{c("norm","basic", "stud", "perc", "bca")} to select the type of bootstrap
confidence interval. See \link[boot:boot.ci]{boot::boot.ci} for more details.}
}
\value{
A list, contains 'descriptive' and 'inferential'
}
\description{
This is a wrapper function to provide adjusted effect estimates and relevant statistics in anchored case (i.e. there
is a common comparator arm in the internal and external trial).
}
\details{
For time-to-event analysis, it is required that input \code{ipd} and \code{pseudo_ipd} to have the following
columns. This function is not sensitive to upper or lower case of letters in column names.
\itemize{
\item USUBJID - character, unique subject ID
\item ARM - character or factor, treatment indicator, column name does not have to be 'ARM'. User specify in
\code{trt_var_ipd} and \code{trt_var_agd}
\item EVENT - numeric, 1 for censored/death, 0 for otherwise
\item TIME - numeric column, observation time of the \code{EVENT}; unit in days
}
}
\examples{
# Anchored example using maic_anchored for time-to-event data

weighted_twt # previously computed result from estimate_weights()
adtte_twt # two arm time-to-event data

pseudo_ipd_twt # digitised time-to-event data

result <- maic_anchored(
  weights_object = weighted_twt,
  ipd = adtte_twt,
  trt_var_ipd = "ARM",
  pseudo_ipd = pseudo_ipd_twt,
  trt_var_agd = "ARM",
  trt_ipd = "A",
  trt_agd = "B",
  trt_common = "C",
  endpoint_name = "Overall Survival",
  endpoint_type = "tte",
  eff_measure = "HR",
  time_scale = "month",
  km_conf_type = "log-log",
)
result$inferential$report_median_surv
result$inferential$report_overall_robustCI
result$inferential$report_overall_bootCI
}
