% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic_unanchored_tte.R
\name{maic_tte_unanchored}
\alias{maic_tte_unanchored}
\title{Unanchored MAIC for time-to-event endpoint}
\usage{
maic_tte_unanchored(
  external = NULL,
  matched_ipd = NULL,
  internal_time_name = NULL,
  internal_event_name = NULL,
  time_scale = "month",
  endpoint_name = "OS",
  transform = "log"
)
}
\arguments{
\item{internal_time_name}{name of the time variable in ipd_matched (for time to event outcome)}

\item{internal_event_name}{name of the event variable in ipd_matched (for time to event outcome)}

\item{time_scale}{a character string, 'year', 'month', 'week' or 'day', time unit of median survival time}

\item{endpoint_name}{a character string, name of the endpoint}

\item{transform}{a character string, pass to \code{\link[survival]{cox.zph}}}

\item{pseudo_ipd}{pseudo comparator IPD. Should be a data frame.
For a time to event outcome, time, status(i.e. event=1), and ARM should be specified
For a response outcome, response and ARM should be specified.}

\item{ipd_matched}{internal IPD data with estimated weights that is returned from \code{\link{estimate_weights}}}

\item{trt}{a character string, name of the interested treatment in internal trial (real IPD)}

\item{trt_ext}{character string, name of the interested comparator in external trial used to subset \code{dat_ext} (pseudo IPD)}
}
\value{
A list of KM plot, analysis table, and diagnostic plot
}
\description{
Unanchored MAIC for time-to-event endpoint
}
\details{
Format requirements for input \code{dat} and \code{dat_ext} are to have the following columns
\itemize{
\item treatment - character or factor column
\item status - logical column, TRUE for censored/death, FALSE for otherwise
\item time - numeric column, observation time of the \code{status}; unit in days
}
}
