% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maicplus.R
\docType{package}
\name{maicplus-package}
\alias{maicplus}
\alias{maicplus-package}
\title{maicplus: R Package Template}
\description{
R package template with GitHub Actions workflows included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hta-pharma/maicplus/}
  \item Report bugs at \url{https://github.com/hta-pharma/maicplus/issues}
}

}
\author{
\strong{Maintainer}: hta-pharma \email{hta-pharma@example.com}

}
\keyword{internal}
