% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maicplus-package.R
\docType{package}
\name{maicplus-package}
\alias{maicplus}
\alias{maicplus-package}
\title{maicplus: Matching Adjusted Indirect Comparison}
\description{
The maicplus package facilitates performing matching adjusted indirect comparison (MAIC) analysis where the endpoint of interest is either time-to-event (e.g. overall survival) or binary (e.g. objective tumor response).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hta-pharma/maicplus/}
  \item Report bugs at \url{https://github.com/hta-pharma/maicplus/issues}
}

}
\author{
\strong{Maintainer}: hta-pharma \email{hta-pharma@example.com}

}
\keyword{internal}
