% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maicplus-package.R
\docType{package}
\name{maicplus-package}
\alias{maicplus}
\alias{maicplus-package}
\title{maicplus: Matching Adjusted Indirect Comparison}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Facilitates performing matching adjusted indirect comparison (MAIC) analysis where the endpoint of interest is either time-to-event (e.g. overall survival) or binary (e.g. objective tumor response). The method is described by Signorovitch et al (2012) \doi{10.1016/j.jval.2012.05.004}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hta-pharma/maicplus/}
  \item \url{https://hta-pharma.github.io/maicplus/}
  \item Report bugs at \url{https://github.com/hta-pharma/maicplus/issues}
}

}
\author{
\strong{Maintainer}: Isaac Gravestock \email{isaac.gravestock@roche.com}

Authors:
\itemize{
  \item Gregory Chen \email{xiangyi.gregory.chen@msd.com}
  \item Michael Seo \email{michael.seo@roche.com}
}

Other contributors:
\itemize{
  \item Miranta Antoniou \email{miranta.antoniou@roche.com} [contributor]
  \item Chrysostomos Kalyvas \email{chrysostomos.kal@gmail.com} [contributor]
  \item MSD, Inc. [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
