% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helper.R
\name{medSurv_makeup}
\alias{medSurv_makeup}
\title{Helper function: makeup to get median survival time from a \code{survival::survfit} object}
\usage{
medSurv_makeup(km_fit, legend = "before matching", time_scale)
}
\arguments{
\item{km_fit}{returned object from \code{survival::survfit}}

\item{legend}{a character string, name used in 'type' column in returned data frame}

\item{time_scale}{a character string, 'year', 'month', 'week' or 'day', time unit of median survival time}
}
\value{
a data frame with a index column 'type', median survival time and confidence interval
}
\description{
Extract and display median survival time with confidence interval
}
