% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{merge_two_data}
\alias{merge_two_data}
\title{Function to merge matched IPD data and pseudo comparator IPD}
\usage{
merge_two_data(
  external = NULL,
  internal = NULL,
  internal_time_name = NULL,
  internal_event_name = NULL,
  internal_response_name = NULL
)
}
\arguments{
\item{internal_time_name}{name of the time variable in ipd_matched (for time to event outcome)}

\item{internal_event_name}{name of the event variable in ipd_matched (for time to event outcome)}

\item{internal_response_name}{name of the response variable in ipd_matched (for binary outcome)}

\item{pseudo_ipd}{data frame of pseudo comparator IPD.
For a time to event outcome, time, event, and ARM should be specified.
For a response outcome, response and ARM should be specified.}

\item{ipd_matched}{internal IPD data with estimated weights that is returned from \link{estimate_weights}}
}
\value{
Merged dataset with time, event, ARM, and weights for time to event data and response, ARM, and weights for binary.
External ARM is assigned to be the reference treatment in the unanchored case. common treatment is assigned to be reference
treatment for the anchored case.
}
\description{
For a survival outcome, we can digitize Kaplan-Meier curves to obtain
pseudo IPD. For a binomial outcome, we can simulate response data based
on the known proportion of responders. This function merges the pseudo
comparator IPD and the internal IPD data. Merged data can be used to
run a regression. For the pseudo IPD, a weight of 1 is assigned.
}
