% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{plot_weights}
\alias{plot_weights}
\title{Plot MAIC weights in a histogram with key statistics in legend}
\usage{
plot_weights(
  wt,
  bin_col = "#6ECEB2",
  vline_col = "#688CE8",
  main_title = "Unscaled Individual Weights"
)
}
\arguments{
\item{wt}{a numeric vector of individual MAIC weights (derived using \code{\link{estimate_weights}})}

\item{bin_col}{a string, color for the bins of histogram}

\item{vline_col}{a string, color for the vertical line in the histogram}

\item{main_title}{a character string, main title of the plot}
}
\value{
a plot of unscaled or scaled weights
}
\description{
Generates a plot given the individuals weights with key summary in top right legend that includes
median weight, effective sample size (ESS), and reduction percentage (what percent ESS is reduced from the
original sample size).
There are two options of weights provided in \code{\link{estimate_weights}}: unscaled or scaled.
Scaled weights are relative to the original unit weights of each individual.
In other words, a scaled weight greater than 1 means that an individual carries more weight in the
re-weighted population than the original data and a scaled weight less than 1 means that an individual carries
less weight in the re-weighted population than the original data.
}
\examples{
load(system.file("extdata","match_res.rda", package = "maicplus", mustWork = TRUE))
wt <- match_res$data$weights
wt_scaled <- match_res$data$scaled_weights
par(mfrow = c(1, 2))
plot_weights(wt, bin_col = "orange", vline_col = "darkblue")
plot_weights(wt_scaled, main_title = "Scaled Individual Weights")

}
