% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{report_table_binary}
\alias{report_table_binary}
\title{helper function: sort out a nice report table to summarize binary analysis results}
\usage{
report_table_binary(
  binobj,
  weighted_result = NULL,
  eff_measure = c("OR", "RD", "RR"),
  tag = NULL
)
}
\arguments{
\item{binobj}{object from glm()}

\item{weighted_result}{object res_AB}

\item{eff_measure}{a string, binary effect measure, could be "OR", "RR", "RD“}

\item{tag}{a string, by default NULL, if specified, an extra 1st column is created in the output}
}
\value{
a data frame with sample size, incidence rate, estimate of binary effect measure with
95\% CI and Wald test of hazard ratio
}
\description{
helper function: sort out a nice report table to summarize binary analysis results
}
