% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helper.R
\name{survfit_makeup}
\alias{survfit_makeup}
\title{helper function: makeup survival::survfit object for km plot}
\usage{
survfit_makeup(km_fit)
}
\arguments{
\item{km_fit}{returned object from \code{survival::survfit}}
}
\value{
a list of data frames, one element per treatment
}
\description{
helper function: makeup survival::survfit object for km plot
}
